Add search and remove features.
The program are now able to search and remove entry from hosts file.

Search function now needs an additional argument which is int array
pointer. The function now will point the supplied pointer to a new pointer
that contains an array of matched line numbers.

The delete function now removes the hosts file instead of overwriting.
This is done for future proofing in case of i decided to support
windows. (Pssst... i kinda dont want to. But who knows.)

Also some minor changes:
* Some segfaults fixed (forgot which one). X(
* Update help message to accomodate new features.
* Update some error messages.